add `appender.py`, `label_merger.py`,  `label_handler.py`,  `score_handler.py`
