+++ src/DAJIN2/core/clustering/__init__.py
@@ -1,4 +1,3 @@
 from DAJIN2.core.clustering.clustering import add_labels
-from DAJIN2.core.clustering.clustering import add_percent
-from DAJIN2.core.clustering.clustering import add_readnum
-from DAJIN2.core.clustering.clustering import update_labels
+from DAJIN2.core.clustering.appender import add_readnum, add_percent
+from DAJIN2.core.clustering.label_handler import update_labels
