+++ src/DAJIN2/core/clustering/make_kmer.py
@@ -11,30 +11,6 @@ def read_json(filepath: Path | str) -> Generator[dict[str, str]]:
             yield json.loads(line)
 
 
-# def generate_mutation_kmers(path_sample: Path | str, mutation_loci: list[set], kmer: int = 5) -> Generator[list[str]]:
-#     midsv_sample = read_json(path_sample)
-#     kmer_n = ",".join(["N"] * kmer)
-#     cssplits = (cs["CSSPLIT"].split(",") for cs in midsv_sample)
-#     for cssplit in cssplits:
-#         mutation_kmers = [kmer_n] * (kmer // 2)
-#         for i in range((kmer // 2), len(cssplit) - (kmer // 2)):
-#             if mutation_loci[i] == set():
-#                 mutation_kmers.append(kmer_n)
-#                 continue
-#             mutations = [cssplit[j][0] for j in range(i - kmer + 1, i + 1)]
-#             if "+" in mutations:
-#                 for j in range(i - kmer + 1, i + 1):
-#                     if cssplit[j][0] == "+":
-#                         cssplit[j] = "+I" + cssplit[j].split("|")[-1]
-#             if all(mutation in mutation_loci[j] for j, mutation in zip(range(i - kmer + 1, i + 1), mutations)):
-#                 kmer = ",".join(cssplit[i - kmer + 1 : i + 1])
-#                 mutation_kmers.append(kmer)
-#             else:
-#                 mutation_kmers.append(kmer_n)
-#         mutation_kmers += [kmer_n] * (kmer // 2)
-#         yield mutation_kmers
-
-
 def generate_mutation_kmers(
     path_sample: Path | str, mutation_loci: list[set[str]], compress_ins: bool = True
 ) -> Generator[list[str]]:
